refactor: add factory suffix to dbLoadSurveys
