+++ src/main/factories/controllers/survey/load-surveys/load-surveys-controller-factory.ts
@@ -1,7 +1,7 @@
 import { LoadSurveysController } from '@/presentation/controllers/survey/load-surveys/load-surveys-controller'
 import { Controller } from '@/presentation/protocols'
 import { makeLogControllerDecorator } from '@/main/factories/decorators/log-controller-decorator-factory'
-import { makeDbLoadSurveys } from '@/main/factories/usecases/survey/load-surveys/db-load-surveys'
+import { makeDbLoadSurveys } from '@/main/factories/usecases/survey/load-surveys/db-load-surveys-factory'
 
 export const makeLoadSurveysController = (): Controller => {
   const controller = new LoadSurveysController(makeDbLoadSurveys())
